+++ src/main/java/ru/practicum/shareit/user/model/User.java
@@ -8,13 +8,13 @@ import lombok.Data;
 @Data
 public class User {
 
-    private Integer id;
+    private Long id;
 
     private String name;
 
     private String email;
 
-    public User(Integer id, String name, String email) {
+    public User(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
